// part9/PatientorBackend/types.ts
// eslint-disable-next-line @typescript-eslint/no-empty-interface
export enum Gender {
    Male = 'male',
    Female = 'female',
    Other = 'other'
}




export interface Patient {
    id: string;
    name: string;
    occupation: string;
    gender: Gender;
    ssn: string;
    dateOfBirth?: string;
    entries: Entry[]
}


export interface Diagnosis {
    code: string;
    name: string;
    latin?: string;
}

export interface NewPatientEntry {
    name: string;
    occupation: string;
    gender: Gender;
    ssn: string;
    dateOfBirth: string;
    entries: Entry[];
}


interface BaseEntry {
    id: string;
    description: string;
    date: string;
    specialist: string;
  
    diagnosisCodes?: Array<Diagnosis['code']>;
  }

  export enum HealthCheckRating {
    "Healthy" = 0,
    "LowRisk" = 1,
    "HighRisk" = 2,
    "CriticalRisk" = 3
  }
  
  interface HealthCheckEntry extends BaseEntry {
    type: "HealthCheck";
    healthCheckRating: HealthCheckRating;
  }
    
  type dischargeobject =  {
    date: string;
    criteria: string;
  }

  interface HospitalEntry extends BaseEntry {
    type: "Hospital";
    discharge: dischargeobject;

  }

  type sickLeave = {
    startDate: string;
    endDate: string;
  }

  interface OccupationalHealthcareEntry extends BaseEntry {
    type: "OccupationalHealthcare";
    employerName: string;
    sickLeave?: sickLeave;
  }

  export type Entry =
  | HospitalEntry
  | OccupationalHealthcareEntry
  | HealthCheckEntry;


// Define special omit for unions
export type UnionOmit<T, K extends string | number | symbol> = T extends unknown ? Omit<T, K> : never;
// Define Entry without the 'id' property
export type EntryWithoutId = UnionOmit<Entry, 'id'>;

export type NonSensitivePatient = Omit<Patient, 'ssn' | 'entries'>;